Reverted long id to int id for programowner table. Separate issue will be created for that

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/trackedentity/TrackedEntityProgramOwner.java
@@ -46,7 +46,7 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
 @JacksonXmlRootElement( localName = "trackedEntityProgramOwner", namespace = DxfNamespaces.DXF_2_0 )
 public class TrackedEntityProgramOwner implements Serializable
 {
-    private long id;
+    private int id;
 
     private TrackedEntityInstance entityInstance;
 
@@ -223,12 +223,12 @@ public class TrackedEntityProgramOwner implements Serializable
         this.createdBy = createdBy;
     }
 
-    public long getId()
+    public int getId()
     {
         return id;
     }
 
-    public void setId( long id )
+    public void setId( int id )
     {
         this.id = id;
     }
